+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/foodies.iml" filepath="$PROJECT_DIR$/.idea/foodies.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/Foodies.iml" filepath="$PROJECT_DIR$/.idea/Foodies.iml" />
     </modules>
   </component>
 </project>
