+++ Foodies/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for djangoProject project.
+Django settings for Foodies project.
 
 Generated by 'django-admin startproject' using Django 5.0.2.
 
@@ -20,7 +20,7 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # See https://docs.djangoproject.com/en/5.0/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = 'django-insecure-%8=^n%s5&7spg68#cy)(t-v-kr5vgeiiz0o&a1j2(!2%z_hcie'
+SECRET_KEY = 'django-insecure-quo3^#z%e$ljrr(6ytt@o$#*5(vy#cwx+-x_jt26m=+y%h-0ft'
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
@@ -37,6 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'djangoProject.apps.DjangoprojectConfig',
 ]
 
 MIDDLEWARE = [
@@ -49,7 +50,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'djangoProject.urls'
+ROOT_URLCONF = 'Foodies.urls'
 
 TEMPLATES = [
     {
@@ -68,7 +69,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'djangoProject.wsgi.application'
+WSGI_APPLICATION = 'Foodies.wsgi.application'
 
 
 # Database
