+++ Foodies/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for djangoProject project.
+WSGI config for Foodies project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'djangoProject.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'Foodies.settings')
 
 application = get_wsgi_application()
