--- models.py
@@ -1,27 +1,0 @@
-from django.db import models
-
-class User(models.Model):
-    username = models.CharField(max_length=50, unique=True)
-    password = models.CharField(max_length=100)
-    email = models.EmailField(unique=True)
-    # Add more fields as needed
-
-class Cuisine(models.Model):
-    name = models.CharField(max_length=100)
-
-    def __str__(self):
-        return self.name
-
-class Restaurant(models.Model):
-    name = models.CharField(max_length=255)
-    description = models.TextField(blank=True, null=True)
-    cuisines = models.ManyToManyField(Cuisine)
-    address = models.CharField(max_length=255)
-    city = models.CharField(max_length=100)
-    state = models.CharField(max_length=100)
-    zip_code = models.CharField(max_length=20)
-    phone_number = models.CharField(max_length=20)
-    website = models.URLField(blank=True, null=True)
-
-    def __str__(self):
-        return self.name
